requirements-dev.txt
# Mocking in unit tests
mock

# Documentation generation
sphinx

# Debugging of database queries, etc.
django-debug-toolbar

# Generation of the model graph used in the docs
django-extensions
